fixed some login bugs and refactored code

+++ Core/src/main/java/ir/sharif/gamein2021/core/repository/TeamRepository.java
@@ -4,6 +4,6 @@ import ir.sharif.gamein2021.core.entity.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface TeamRepository extends JpaRepository<Team, Long> {
+public interface TeamRepository extends JpaRepository<Team, Integer> {
 
 }
